backend/requirements.txt
appnope==0.1.0
astroid==2.3.2
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.7
django-cors-headers==3.1.1
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
entrypoints==0.3
gunicorn==20.0.4
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.1.1
jupyter-client==5.3.4
jupyter-core==4.6.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
nose==1.3.7
notebook==6.0.1
numpy==1.17.3
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pipenv==2018.11.26
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.4.2
pylint==2.4.3
pyrsistent==0.15.5
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.3
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.22.0
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.3.0
tabulate==0.8.6
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.6
virtualenv==16.7.9
virtualenv-clone==0.5.3
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==5.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
